requirements.txt
Babel==2.7.0		
async-timeout==3.0.1	
asyncio==3.4.3	
attrs==19.1.0	
certifi==2019.3.9	
chardet==3.0.4	
idna==2.8	
multidict==4.5.2	
pip==19.0.3	
pyTelegramBotAPI==3.6.6			
pytz==2019.2	
requests==2.7.0	
setuptools==40.8.0	
six==1.12.0	
urllib3==1.25.3	
websockets==6.0	
yarl==1.3.0
schedule==0.6.0
psycopg2-binary==2.8.3	
